AstOptimizer

Documentation

optimizeAst :: [ScopeMb] -> [Ast] -> Bool -> [Either AstError AstOptimised]

Optimize a list of Ast and check for invalid operation: optimization is taking place when operation have the same result no matter what for exemple `3 + 3`, when a forbidden operation is taking place, the optimization results in an AstError, when the optimization was a success it results in an AstOptimised. Takes the stack (`[ScopeMb]`), a list of Ast, a boolean to indicate whether the optimization take place insinde a function and returns the list of Either AstError or AstOptimised

fromOpti :: AstOptimised -> Ast

Get the Ast contained in a AstOptimised

data AstError

Constructors

Error String Ast

data AstOptimised

Constructors

Result Ast
The Ast after optimization

Warning String Ast
When the optimization throw a warning contains the warining message and the Ast post optimization that caused it