Parse

Documentation

data Parser a

Constructors

Parser

Fields

runParser :: Position -> String -> Either String (a, String, Position)

parseChar :: Char -> Parser Char

Parse a character c Takes the character that need to be parsed Returns a data Parser that contain the character and the rest of the string

parseOr :: Parser a -> Parser a -> Parser a

Parse with the first or the second parser Takes two parsers Returns either the first parser or the second parser

parseAnd :: Parser a -> Parser b -> Parser (a, b)

Parse with the first and the second parser Takes two parsers Returns either the first parser then use result for the second parser

parseAndWith :: (a -> b -> c) -> Parser a -> Parser b -> Parser c

Parse with function after the two parsers Takes two parsers and a fh (x _ -> x) unction Returns the result of the function with the result of the parseAnd

parseMany :: Parser a -> Parser [a]

Parse with a parser Takes a parser Returns the application of the parser (if nothing, returns an empty list)

parseSome :: Parser a -> Parser [a]

Parse with a parser Takes a parser Returns the application of the parser at least one time or Returns Nothing

parseUInt :: Parser Int

Return a data Parser that parse a UInt

parseInt :: Parser Int

Return a data Parser that parse a Int

parseAnyChar :: String -> Parser Char

Parse any character from a String Return a Parser that parse every character from a String

parseList :: Parser a -> String -> String -> Parser [a]

Parse a list of element Return a Parser of list element that start with a '(' and end with a ')'

parseString :: Parser String

Return a data Parser that parse a String

parseSign :: Parser Char

Return a data Parser that parse a - or +

parseDigit :: Parser Char

Return a data Parser that parse a digit

parseBool :: Parser Ast

Return a Parser that parse a Bool (#f or #t)

parseAst :: Parser Ast

Return a Parser that parse a SExpr

parseElem :: Parser a -> Parser a

Parse with a parser and, if possible with a space Return a Parser that parse element with the given parser and, if possible with multiple space

parseValue :: Parser Ast

Return a data Parser that parse a Int as a Value

parseLobster :: Parser [Ast]

parseAnyString :: String -> Parser String

Parse a specific String

parseCmpString :: String -> Parser String

parseDefineValue :: Parser Ast

parseProduct :: Parser Ast

parseSum :: Parser Ast

parseExpr :: Parser Ast

parseTrue :: Parser Bool

Return a PArser that parse a True (in lisp -> #t)

parseFalse :: Parser Bool

Return a PArser that parse a True (in lisp -> #f)

parseAstString :: Parser Ast

Return a data Parser that parse a String as a Symbol

parseWhiteSpace :: Parser [Char]

errorParsing :: (Int, Int) -> String

parseDefineFn :: Parser Ast

parseLambda :: Parser Ast

parseCond :: Parser Ast

parseFunctionValue :: Parser Ast

parseBracket :: Parser Ast

parseComment :: Parser Char