Scope

Documentation

data ScopeMb

Structure representing a member in a scope. Can be a ScopeBegin to mark the beginning of a scope or a Variable containing its name as a Ast and its value as an Ast

Constructors

ScopeBegin Int

Variable String Ast Int

beginScope :: [ScopeMb] -> [ScopeMb]

Begin a new scope by adding a ScopeBegin to the stack.

clearScope :: [ScopeMb] -> [ScopeMb]

Clear the current scope

addVarToScope :: [ScopeMb] -> String -> Ast -> [ScopeMb]

Add a Variable to the stack with its name as a Ast and its value by an Ast

getVarInScope :: [ScopeMb] -> String -> Maybe Ast

Get the value contained in the variable given by name as a Ast, return Nothing if the variable don't exist or Just its value

addVarsToScope :: [ScopeMb] -> [String] -> [Ast] -> [ScopeMb]

Add multiple variables to the stack with their names as a Ast and their values as an Ast

updateVar :: [ScopeMb] -> String -> Ast -> [ScopeMb]

Update the given variable, don't do anything if it doesn't exist