Stack

Documentation

top :: [a] -> Maybe a

Return the top of the stack

push :: [a] -> a -> [a]

Push a value onto the stack, return the new stack

pop :: [a] -> (Maybe a, [a])

Remove the top of the stack, return a tuple containing the old top and the new stack

seek :: (a -> Bool) -> [a] -> Maybe a

Return the first element of the stack validating the function given as parameter

size :: [a] -> Int

Return the size of the stack

clear :: [a] -> [a]

Clear the stack by return an empty stack